requirements.txt
torch
mace-torch
ase
fastapi
uvicorn
pydantic
python-multipart
rich
